# Improve find_interval with binary search