The following benchmark and improved implementation were developed with the help of LLM ClaudeAI.

I used a benchmark to analyze the overall performance of the library (see benchmark at the end).

The initial results showed a performance issue in the "Finding interval" part:

```
=== Ephem Performance Benchmark ===
SPK File: tmp/de421.bsp
Center: 0 (Solar System Barycenter)
Target: 5 (Jupiter)
Time: 2460676.5 (Julian Date)

=== Benchmarking Initialization ===
Opening SPK file: 0.2859999658539891 ms
Getting segment: 0.0029999064281582832 ms

=== Benchmarking Single Computation ===
Loading segment data: 2.440999960526824 ms
Converting time: 0.00800006091594696 ms
Finding interval: 0.35399990156292915 ms
Evaluating position polynomial: 0.04700000863522291 ms
Evaluating velocity polynomial: 0.0540000619366765 ms
Full compute_and_differentiate: 0.4290000069886446 ms

=== Benchmarking Multiple Iterations ===
Average compute time (10000 iterations): 0.40389119983883576 ms

=== Memory Usage Analysis ===
Memory usage before computation: 25136 KB
Memory usage after computation: 25136 KB
Memory increase: 0 KB

=== File Size Analysis ===
SPK file size: 16788480 bytes (16.0107421875 MB)
Number of segments: 15

=== Binary Reading Analysis ===
Average record read time: 0.0056999968364834785 ms
Reading segment array data: 0.36399997770786285 ms
```

After implementing a binary search instead of a linear search, the results were very positive:

```
=== Ephem Performance Benchmark ===
SPK File: tmp/de421.bsp
Center: 0 (Solar System Barycenter)
Target: 5 (Jupiter)
Time: 2460676.5 (Julian Date)

=== Benchmarking Initialization ===
Opening SPK file: 1.2410000199452043 ms
Getting segment: 0.00400003045797348 ms

=== Benchmarking Single Computation ===
Loading segment data: 2.658999990671873 ms
Converting time: 0.007999944500625134 ms
Finding interval: 0.0069999368861317635 ms
Evaluating position polynomial: 0.056999968364834785 ms
Evaluating velocity polynomial: 0.054999953135848045 ms
Full compute_and_differentiate: 0.0849999487400055 ms

=== Benchmarking Multiple Iterations ===
Average compute time (10000 iterations): 0.07585140033625067 ms

=== Memory Usage Analysis ===
Memory usage before computation: 41888 KB
Memory usage after computation: 41888 KB
Memory increase: 0 KB

=== File Size Analysis ===
SPK file size: 16788480 bytes (16.0107421875 MB)
Number of segments: 15

=== Binary Reading Analysis ===
Average record read time: 0.007300009019672871 ms
Reading segment array data: 0.41800003964453936 ms
```

- Finding interval: 98% faster (0.354ms → 0.007ms)
- Full computation: 80% faster (0.429ms → 0.085ms)
- Average compute time: 81% faster (0.404ms → 0.076ms)

### Benchmark

```rb
# frozen_string_literal: true

require "benchmark"
require "ephem"

# File to benchmark
SPK_FILE = "tmp/de421.bsp"
CENTER = 0  # Solar System Barycenter
TARGET = 5  # Jupiter
TIME = 2460676.5

# Number of iterations for averaging
ITERATIONS = 10_000

def benchmark_init
  puts "\n=== Benchmarking Initialization ==="

  time = Benchmark.measure do
    @spk = Ephem::SPK.open(SPK_FILE)
  end
  puts "Opening SPK file: #{time.real * 1000} ms"

  time = Benchmark.measure do
    @segment = @spk[CENTER, TARGET]
  end
  puts "Getting segment: #{time.real * 1000} ms"
end

def benchmark_compute_once
  puts "\n=== Benchmarking Single Computation ==="

  # Force data reload to ensure accurate measurement
  @segment.clear_data

  # Measure data loading separately
  data_time = Benchmark.measure do
    @segment.send(:load_data)
  end
  puts "Loading segment data: #{data_time.real * 1000} ms"

  # Measure time conversion
  convert_time = Benchmark.measure do
    @tdb_seconds = @segment.send(:convert_to_seconds, TIME, 0.0)
  end
  puts "Converting time: #{convert_time.real * 1000} ms"

  # Measure interval finding
  interval_time = Benchmark.measure do
    @interval = @segment.send(:find_interval, @tdb_seconds)
  end
  puts "Finding interval: #{interval_time.real * 1000} ms"

  # Get normalized time and coefficients
  @normalized_time = @segment.send(:compute_normalized_time, @tdb_seconds, @interval)
  @coeffs = @segment.instance_variable_get(:@coefficients)[true, true, @interval]

  # Measure position polynomial evaluation
  pos_time = Benchmark.measure do
    @position = Ephem::Computation::ChebyshevPolynomial.new(
      coefficients: @coeffs,
      normalized_time: @normalized_time
    ).evaluate
  end
  puts "Evaluating position polynomial: #{pos_time.real * 1000} ms"

  # Measure velocity polynomial evaluation
  vel_time = Benchmark.measure do
    @velocity = Ephem::Computation::ChebyshevPolynomial.new(
      coefficients: @coeffs,
      normalized_time: @normalized_time,
      radius: @segment.instance_variable_get(:@radii)[@interval]
    ).evaluate_derivative
  end
  puts "Evaluating velocity polynomial: #{vel_time.real * 1000} ms"

  # Measure entire computation
  full_time = Benchmark.measure do
    @state = @segment.compute_and_differentiate(TIME)
  end
  puts "Full compute_and_differentiate: #{full_time.real * 1000} ms"
end

def benchmark_multiple_iterations
  puts "\n=== Benchmarking Multiple Iterations ==="

  # Cache is warm now, measure with cached data
  times = []
  ITERATIONS.times do |i|
    time = Benchmark.measure do
      @state = @segment.compute_and_differentiate(TIME + i * 0.1)
    end
    times << time.real * 1000
  end

  avg_time = times.sum / times.size
  puts "Average compute time (#{ITERATIONS} iterations): #{avg_time} ms"
end

def profile_memory_usage
  puts "\n=== Memory Usage Analysis ==="
  memory_before = `ps -o rss= -p #{Process.pid}`.to_i
  puts "Memory usage before computation: #{memory_before} KB"

  # Run a large number of computations
  50.times do |i|
    @segment.compute_and_differentiate(TIME + i * 0.1)
  end

  memory_after = `ps -o rss= -p #{Process.pid}`.to_i
  puts "Memory usage after computation: #{memory_after} KB"
  puts "Memory increase: #{memory_after - memory_before} KB"
end

def file_size_analysis
  puts "\n=== File Size Analysis ==="
  size = File.size(SPK_FILE)
  puts "SPK file size: #{size} bytes (#{size / 1024 / 1024.0} MB)"

  # Get segment count and sizes
  segment_count = @spk.segments.size
  puts "Number of segments: #{segment_count}"
end

def analyze_binary_reading
  puts "\n=== Binary Reading Analysis ==="

  # Get the binary reader
  binary_reader = @spk.instance_variable_get(:@daf).instance_variable_get(:@binary_reader)

  # Measure random record reads
  times = []
  10.times do |i|
    time = Benchmark.measure do
      binary_reader.read_record(i + 1)
    end
    times << time.real * 1000
  end

  avg_time = times.sum / times.size
  puts "Average record read time: #{avg_time} ms"

  # Test array reading
  start_i = @segment.instance_variable_get(:@start_i)
  end_i = @segment.instance_variable_get(:@end_i)

  time = Benchmark.measure do
    @spk.instance_variable_get(:@daf).read_array(start_i, end_i - 4)
  end
  puts "Reading segment array data: #{time.real * 1000} ms"
end

def run_benchmarks
  puts "=== Ephem Performance Benchmark ==="
  puts "SPK File: #{SPK_FILE}"
  puts "Center: #{CENTER} (Solar System Barycenter)"
  puts "Target: #{TARGET} (Jupiter)"
  puts "Time: #{TIME} (Julian Date)"

  # Run the benchmarks
  benchmark_init
  benchmark_compute_once
  benchmark_multiple_iterations
  profile_memory_usage
  file_size_analysis
  analyze_binary_reading

  # Clean up
  @spk.close
end

# Run all benchmarks
run_benchmarks
```